Added even more checks.

+++ src/createcompendia/publications.py
@@ -118,8 +118,10 @@ def verify_pubmed_downloads(pubmed_directories,
                 while not verify_pubmed_download_against_md5(file_path, file_path + '.md5'):
                     pubmed_url_base = pubmed_base + subdir + '/'
                     logging.warning(f"Re-downloading {file_path} from HTTPS URL {pubmed_url_base}{filename}.")
-                    os.remove(file_path)
-                    os.remove(md5_file_path)
+                    if os.path.exists(file_path):
+                        os.remove(file_path)
+                    if os.path.exists(md5_file_path):
+                        os.remove(md5_file_path)
                     pull_via_wget(pubmed_url_base, filename, decompress=False, subpath='PubMed/' + subdir)
                     pull_via_wget(pubmed_url_base, md5_filename, decompress=False, subpath='PubMed/' + subdir)
